+++ align/pnr/cmdline.py
