+++ bin/pnr_compiler.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from align.pnr.pnr_compiler import cmdline
+from align.pnr.cmdline import cmdline
 import sys
 
 if __name__ == '__main__':
