+++ tests/pnr/test_pnr_cmdline.py
@@ -3,7 +3,7 @@ import pathlib
 import io
 import pytest
 
-from align.pnr import pnr_compiler
+from align.pnr.cmdline import cmdline
 
 mydir = pathlib.Path(__file__).resolve().parent
 
@@ -15,7 +15,7 @@ def test_verilog():
 
     argv = [ 'pnr_compiler.py', str(d), f'{nm}.lef', f'{nm}.v', f'{nm}.map', 'layers.json', nm, '1', '0']
 
-    pnr_compiler.cmdline( argv)
+    cmdline( argv)
 
 def test_verilog_json():
     
@@ -25,4 +25,4 @@ def test_verilog_json():
 
     argv = [ 'pnr_compiler.py', str(d), f'{nm}.lef', f'{nm}.verilog.json', f'{nm}.map', 'layers.json', nm, '1', '0']
 
-    pnr_compiler.cmdline( argv)
+    cmdline( argv)
